many improvements to world tracking

+++ src/alloc.c
@@ -758,7 +758,7 @@ JL_DLLEXPORT jl_sym_t *jl_symbol_lookup(const char *str)
     return symtab_lookup(&symtab, str, strlen(str), NULL);
 }
 
-JL_DLLEXPORT jl_sym_t *jl_symbol_n(const char *str, int32_t len)
+JL_DLLEXPORT jl_sym_t *jl_symbol_n(const char *str, size_t len)
 {
     if (memchr(str, 0, len))
         jl_exceptionf(jl_argumenterror_type, "Symbol name may not contain \\0");
